+++ models/points_test.go
@@ -1976,6 +1976,12 @@ func TestParseKeyEmpty(t *testing.T) {
 	}
 }
 
+func TestParseKeyMissingValue(t *testing.T) {
+	if _, _, err := models.ParseKey([]byte("cpu,foo ")); err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+}
+
 func TestPoint_FieldIterator_Simple(t *testing.T) {
 
 	p, err := models.ParsePoints([]byte(`m v=42i,f=42 36`))